ajuste de formatação

+++ src/cliente.c
@@ -1,6 +1,6 @@
 #include "cliente.h"
 
-Cliente* criar_cliente(char* nome, int cpf, int prioridade, int qtd_produtos) {
+Cliente* criar_cliente(char* nome, long cpf, int prioridade, int qtd_produtos) {
     Cliente* cliente = (Cliente*) malloc(sizeof(Cliente));
     if (cliente == NULL) {
         printf("Erro ao alocar memória para o cliente.\n");
